+++ AtmelStudio/Drafts/StateMachineGarageDoor/StateMachineGarageDoor/settings.h
@@ -1,31 +1,40 @@
 #ifndef settings_H
 #define settings_H
 
-/* set up LEDs
-opening - green LED blinking
-closing - white LED blinking
-open - green LED on
-closed - white LED on
-alarm - red LED on
+//Define input and output registers and ports
+#define INPUT_REG			DDRB	//Register for inputs
+#define OUTPUT_REG			DDRC	//Register for outputs
+#define INPUT_PORT			PORTB	//Port for inputs
+#define OUTPUT_PORT			PORTC	//Port for outputs
+#define INPUT_PIN			PINB	//Pin for inputs
+#define OUTPUT_PIN			PORTC	//Pin for outputs
+
+//Buttons		 					- ################ INPUTS ################
+//Pins for push buttons to trigger opening or closing the door
+#define OPEN_BTN_PIN		PB0		//Brown wire
+#define CLOSE_BTN_PIN		PB1		//Green wire
+
+//Pins for switches that signal that door was fully closed or fully open
+#define OPEN_SWITCH_PIN		PB2		//Red wire
+#define CLOSE_SWITCH_PIN	PB3		//White wire
+
+//Emergency push button to stop the motor running
+//#define MOTOR_STOP_PIN		PB4
+
+/*LEDs, motor, lock					- ################ OUTPUTS ################
+opening, closing	- green and white LEDs on
+open				- green LED on
+closed				- white LED on
+alarm				- red LED on
+RF signal			- blue LED on
 */
 #define OPEN_LED_PIN		PC0		//Port for Open LED
 #define CLOSE_LED_PIN		PC1		//Port for Close LED
 #define ALARM_LED_PIN		PC2		//Port for Alarm LED
 //#define RF_LED_PIN			PC3		//Port for RF LED
-
-//Button definition
-//On which pins are push buttons to trigger opening or closing the door
-#define OPEN_BTN_PIN		PD0		//Brown wire
-#define CLOSE_BTN_PIN		PD1		//Green wire
-
-/*Switch (touchguard) definition
-On  which pins are switches that signal that door was fully closed or fully open
-*/
-#define OPEN_SWITCH_PIN		PD2		//Red wire
-#define CLOSE_SWITCH_PIN	PD3		//White wire
-
-//Emergency push button to stop the motor running
-//#define MOTOR_STOP_PIN		PD4
+#define MOTOR_IN1_PIN		PC4		//To motor's IN1
+#define MOTOR_IN2_PIN		PC5		//To motor's IN2
+//#define LOCK_PIN			PC6		//Pin for electro magnetic lock (relay for it)
 
 //States definition
 //Define all states of the machine